+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to Semantic Versioning.
 
 ## [Unreleased]
+### Added
+- Add AlphaPointClient class that calls the WebSocket client.
+- Add response classes for the products, instruments, and miscellaneous sections.
 
 ## [1.0.0-alpha2] - 2018-10-10
 ### Added